Modify setup scripts to accept database parameters
allowing an optional argument to specify the database connection
parameters allows different connection methods to be used. additionally,
along with some script refactoring, it allows the database to be setup
separately from the abot server.

to support this, cmd/setup.sh was refactored into the following scripts:
- cmd/setup.sh
- cmd/dbsetup.sh
- cmd/helper_functions.sh

cmd/setup.sh now accepts an optional argument for the database
connection parameters. if no argument is given, it assumes sane
defaults. in both cases, the database parameters are saved in abot.env
under ABOT_DATABASE_URL. during setup, if the psql binary is
unavailable, the script will skip the database setup. it assumes it was
handled externally. otherwise, cmd/dbsetup.sh is called with the same
connection argument (if any) that this script was called with. beyond
that, code was refactored into cmd/dbsetup.sh and
cmd/helper_functions.sh and some general cleanup performed.

cmd/dbsetup.sh contains all the logic for setting up the abot databases.
it also accepts the same optional argument as cmd/setup.sh. this allows
the database to be setup on a different machine than the abot server.
this script now encapsulates the dependency on the psql binary.

cmd/helper_functions.sh contains functions that help with running setup
commands, pretty printing messages, and database parameter parsing. this
allows code re-use between cmd/setup.sh and cmd/dbsetup.sh, which may be
called independently.

the following changes were made to the go code-base:

abot.go:loadEnvVars() is now in core/boot.go:LoadEnvVars(). since the
database connection relies on the ABOT_DATABASE_URL environment
variable, this function needs to be available for plugins/external
packages.

core/boot.go:ConnectDB() now uses the ABOT_DATABASE_URL environment
variable exclusively. if ABOT_DATABASE_URL is empty, it applies
defaults.

shared/plugin/plugin.go:connectDB() is gone. now it uses the code in
core/boot.go:ConnectDB()
